+++ packages/datepicker/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.15](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-datepicker@0.9.14...@equinor/fusion-react-datepicker@0.9.15) (2021-11-23)
+
+**Note:** Version bump only for package @equinor/fusion-react-datepicker
+
+
+
+
+
+## [0.9.14](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-datepicker@0.9.13...@equinor/fusion-react-datepicker@0.9.14) (2021-11-22)
+
+
+### Bug Fixes
+
+* style rules ([7126199](https://github.com/equinor/fusion-react-components/commit/7126199dfb224b8bf8ac7e72874750ad3b88ca60))
+
+
+
+
+
+## 0.9.13 (2021-11-22)
+
+**Note:** Version bump only for package @equinor/fusion-react-datepicker
+
+
+
+
+
 ## 0.9.12 (2021-11-08)
 
 **Note:** Version bump only for package @equinor/fusion-react-datepicker