+++ packages/filter/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.6](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-filter@1.2.5...@equinor/fusion-react-filter@1.2.6) (2021-11-23)
+
+**Note:** Version bump only for package @equinor/fusion-react-filter
+
+
+
+
+
+## [1.2.5](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-filter@1.2.4...@equinor/fusion-react-filter@1.2.5) (2021-11-22)
+
+**Note:** Version bump only for package @equinor/fusion-react-filter
+
+
+
+
+
+## [1.2.4](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-filter@1.2.3...@equinor/fusion-react-filter@1.2.4) (2021-11-22)
+
+**Note:** Version bump only for package @equinor/fusion-react-filter
+
+
+
+
+
 ## [1.2.3](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-filter@1.2.2...@equinor/fusion-react-filter@1.2.3) (2021-11-15)
 
 **Note:** Version bump only for package @equinor/fusion-react-filter