+++ packages/switch/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.4.5 (2021-11-23)
+
+**Note:** Version bump only for package @equinor/fusion-react-switch
+
+
+
+
+
+## 0.4.4 (2021-11-22)
+
+
+### Bug Fixes
+
+* **switch:** expose disable prop ([1efec30](https://github.com/equinor/fusion-react-components/commit/1efec304edee3a0bd4a72cd4f589e38f93995558))
+
+
+
+
+
 ## [0.4.3](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-switch@0.4.2...@equinor/fusion-react-switch@0.4.3) (2021-11-02)
 
 **Note:** Version bump only for package @equinor/fusion-react-switch